routeTable alpha release
undaqTools.routetable now has RouteTable object that can parse the NADS
RouteTable format. It also provides some additional functionality for
packing and unpacking device data and input data.

In the scripts folder are two new scripts. routetableclient.py is an
echo client for testing the RouteTable functionality.
routeTableEmulator.py is intended as a developer tool. It emulates the
output functionality of RouteTable offline.